feat: capture cancer subset in DO (#251)
We're able to identify whether a disease is or relates to a cancer in
some other sources. This adds the same property to DO terms.

* Use case is pretty narrow -- I'd like to slightly restrict exploration
of the "inexact xrefs" issue (#244) to cancers for now, and I figured it
made sense to just add the code to the library itself instead of keeping
it in my analysis script